Add logging interceptor
This patch adds a logging interceptor that writes to the log the details
of requests and responses. It is configurable via the `--log-headers`
and `--log-bodies` command line flags. By default is only enabled when
the log level is set to debug (with `--log-level=debug`) and only writes
the names of the methods, but not the request or response bodies.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>