Add database change detector
This patch adds a component that detects changes in the database. The
database has a new `changes` table to store the changes and a set of
triggers that write to that table and notifies the server when there are
new changes to process. The server has new _change detector_ componet
that watches that table and processes the changes. Currently that
processing just writes the changes to the log, but in the future it will
publish them to a Kafka topic.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>