Add DAO events
This patch modifies the generic DAO so that it triggers events when
objects are created, updated or deleted.

In later changes these events will be used by the API servers to
generate events when objects are created, modified or deleted. Those API
events will then be published using the PostgreSQL's notify command, and
consumed using the PostgreSQL'S listen command by the events server,
which will then forward them to the clients that are watching.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>