Use builder patter to create DAO requests
Currently the operators of the generic DAO require the user to
explicitly create a request struct containg the parameters. As those
structs are generic and Go can't infer their type, the result is quite
verbose because the type has to be explcitly indicated. For example, to
call the `List` operation:

```go
response, err := myDao.List(ctx, dao.ListRequest[*fulfillmentv1.Cluster]{
        Offset: 10,
	Limit:  100,
	Filter: "!has(this.metadata.deletion_timestamp)",
})
if err != nil {
        ...
}
```

To make that a bit less verbose this patch changes the DAO so that it
creates those structs internally, and the user only has to populate the
parameters. For example, the same `List` operation will now be like
this:

```go
response, err := myDao.List().
	SetOffset(10).
	SetLimit(100).
	SetFilter("!has(this.metadata.deletion_timestamp").
	Send(ctx)
if err != nil {
        ...
}
```

This parameter passing style is applied to all the operations of the
DAO.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>