Add keep alive to gRPC client to prevent disconnections
This change enables the gRPC keep alive functionality in an attempt to
avoid connections being forcibly closed by proxies, particularly Envoy
in our deployment environment. We have been experiencing intermittent
connection issues that we haven't been able to fully identify, but they
appear to be related to idle connection timeouts at the proxy layer.

The implementation adds a new `--grpc-keep-alive` command line flag
allowing us to specify an interval for periodic keep alive events. When
configured, the client will periodically ping the server to verify that
the connection is alive. This should significantly reduce the chances of
proxy-initiated connection drops during periods of low event activity.

In addition, in the Envoy configuration a new route is added
specifically for the `Watch` method of the events service. This route is
configured so that the total and idle timeouts are disabled.

Related: https://github.com/innabox/fulfillment-service/issues/111
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>