Return not found error when deleting non-existent objects
This change introduces a new `ErrNotFound` error type in the DAO layer
that is returned when attempting to delete an object that doesn't exist.
Previously, the `Delete` method would silently succeed even when no object
was found, which made it difficult for callers to distinguish between
successful deletion and attempts to delete non-existent objects.

The generic server has been updated to translate the `ErrNotFound` error
into a gRPC `NotFound` status code, ensuring that API clients receive the
appropriate HTTP 404 response. Tests have been added at both the DAO and
server layers to verify that the error is correctly returned and
propagated when deleting objects that don't exist or are not visible due
to tenant filtering.

Related: https://github.com/innabox/fulfillment-service/issues/138
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>