Use default integration tests CA in Keycloak
This patch changes the Helm chart that installs the Keycloak instance
used by integration tests so that it uses the `default-ca` cert-manager
issuer created by the integration tests infrastructure instead of its
own issuer. This simplifies the chart and will eventually simplify use
of trust bundles in other parts of the integration tests.

The patch also makes the issuer configurable, in case someone wants to
use the chart in some other environment. For example, to issuer
`my-issuer`:

```
$ helm install keycloak charts/keycloak \
--namespace keycloak \
--create-namespace \
--set issuerRef.name=my-issuer \
--wait
```

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>