Configure Authorino to trust Keycloak
This patch changes the configuration of Authorino so that it trusts the
CA certificate of Keycloak. It also adds an alternative authentication
configuration that uses the JSON web tokens issued by Keycloak. These
tokens are accepted for authentication, but not yet accepted for
authorization. That will be done in separate patches.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>