Externalize CA
Currently the service uses a _cert-manager_ issuer that is created as
part of the deployment of the service. In general users will already
have their own issuer which they will want to re-use. To make that
simpler this patch changes the service so that it uses a `default-ca`
cluster-wide issuer that the user should create before deploying the
service. An example of that issuer is included in the Helm chart inside
the `charts/ca` directory.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>