Add JWT tenancy logic with delegation pattern
The default tenancy logic now uses a delegation pattern to support both
service accounts and users autenticate with JSON web tokens. It checks
the source field in the subject to determine which delegate to use,
maintaining backwards compatibility with the existing service account
logic while adding support for JSON web tokens.

The JSON web token tenancy logic is intentionally simple, extracting
tenants from the groups claim in the token. Assigned tenants are the
user's groups, and visible tenants include those groups plus the shared
tenant. This assumes the presence of username and groups claims in the
token, which can be made more sophisticated in later patches if needed.

To support this new tenancy logic this patch also modifies the Keycloak
realm used in the integration tests so that the JSON web tokens that it
issues (for the `fulfillment-cli` client) will contain a claim named
`username` containing the user name and a claim named `groups`
containing the names of the groups that the user belongs to.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>