WIP: Add optional Helm chart for BCM inventory synchronization
This change introduces a new optional Helm chart named `bcm-sync` that
can be deployed independently to synchronize inventory data from _NVIDIA
Base Command Manager_ (BCM) into the fulfillment service.

The synchronizer periodically queries the BCM API to retrieve device
information and updates the fulfillment service database accordingly. The
chart includes a deployment manifest that runs the synchronizer as a
containerized workload, along with a service account for authentication
purposes. The synchronizer has been granted the necessary permissions in the
service authorization configuration to access the private gRPC API.

The BCM synchronizer uses device UUIDs as host identifiers, and BCM
device names as host names. Additionally, it populates the `bmc` and
`rack` field that were added to the `HostSpec` protocol buffer
definition to store the information that will allow in the future the
selection and power management of hosts for the installation of
clusters.

The chart configuration supports customization of the BCM API endpoint URL,
authentication credentials via Kubernetes secrets, synchronization interval,
logging levels, and gRPC server connection details.

To install the chart, create a Kubernetes secret containing the BCM
credentials and then deploy the chart using _Helm_. For example:

```bash
$ kubectl create secret generic bcm-credentials \
--from-file=tls.crt=admin.pem \
--from-file=tls.key=admin.key \
--from-file=ca.crt=cacert.pem \
-n innabox

$ helm upgrade --install bcm-sync charts/bcm-sync \
--namespace innabox \
--set images.service=ghcr.io/innabox/fulfillment-service:main \
--set certs.caBundle.configMap=ca-bundle \
--set bcm.url=https://... \
--set bcm.credentialsSecret=bcm-credentials \
--set bcm.syncInterval=10s \
--wait
```

Once deployed, the synchronizer will periodically fetch device information
from BCM and create or update host records in the fulfillment service. The
synchronized hosts will include rack location information extracted from
BCM, as shown in the following example output:

```bash
$ fulfillment-cli get private.v1.Host
ID                                    NAME     POWER STATE  RACK
024c8f3a-1422-46d6-b5ea-fdac43178dd6  node002  UNSPECIFIED  1
285fc71d-c0f7-421e-bc31-870545943987  node001  UNSPECIFIED  1
```

The power management details will also be available:

```bash
$ fulfillment-cli get private.v1.Host -o yaml
- '@type': type.googleapis.com/private.v1.Host
  id: 024c8f3a-1422-46d6-b5ea-fdac43178dd6
  metadata:
    creation_timestamp: "2025-11-12T21:33:02.173195Z"
    creators:
      - system
    name: node002
    tenants:
      - shared
  spec:
    bmc:
      password: ...
      url: https://192.168.101.22
      user: ...
    rack: "1"
- '@type': type.googleapis.com/private.v1.Host
  id: 285fc71d-c0f7-421e-bc31-870545943987
  metadata:
    creation_timestamp: "2025-11-12T21:33:02.118645Z"
    creators:
      - system
    name: node001
    tenants:
      - shared
  spec:
    bmc:
      password: ...
      url: https://192.168.101.21
      user: ...
    rack: "1"
```

Noe that currently these new attributes (`bmc` and `rack`) are only
visible using the private API, as they are only for internal use.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>

Fix translation of `in` with empty list

Currently we translate a CEL expression like `this.id in []` into SQL
like `where id in ()`. But that is a syntax error in PostgreSQL. To
avoid that this patch changes the translator so that it will translate
it into `false`.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>