Use Redfish API to find system identifiers
BCM can have for each host a BMC network, which includes the IP address
and a type like `rf0' for _Redfish_, `ilo0` for HP iLO, etc. But it
doesn't store the complete URL or system identifier that is needed by
_metal3_ to perform its operation. This patch changes the BCM
synchronizer so that it uses the _Redfish_ API to connect to the IP
address and try to find the system identifier.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>