Create DNS records
This patch changes the cluster reconciler so that it creates the minimum
set of DNS records to make the installation work:

- `api.my-cluster.my-domain.com`
- `api-int.my-cluster.my-domain.com`
- `*.apps.my-cluster.my-domain.com`

The API domains point to the VIP of the management cluster, because we
are using the `NodePort` publishing strategy.

The ingress domains uses DNS round robin to the worker nodes of the
hosted cluster.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>