Add `metadata.tenants`
This fields adds a new `metadata.tenants` field to objects. It will
contain the idenfitifers of the tenants that are assigned to the object.
Users will only get the subset of tenants that can see. For example, if
the object has tenants `marketing` and `engineering` but the user only
has visibility of the `marketing` tenant, then the field will only
contain `marketing`.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>