Add support for explicitly setting the assigned tenant
Currently when a user creates an object the system automatically assigns
to the object the tenants determined by the tenancy logic, which usually
are all the tenants of the user. But in some situations users will want
to explicitly select one of the tenants that they belong to. This patch
adds suport for that: the user will be able to set the value for the
`metadata.tenants` field, and that will be respected by the server as
log as the specified tenants are visibile to the user and don't
contradict the decision of the tenancy logic.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>