Cache visibile and assigned tenants during request
Currently the calculation of assigned and visibile tenants is done each
time that it is needed. This means, for example, that for a request to
retrieve 100 clusters it will this calculation will be done at least 100
times. To avoid that this patch changes the generic DAO so that will do
this calculation only when needed, and will keep it cached for the
duration of the request.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>