Kick controller sync loop before watch
The code that controls the sync and watch loops of the reconcilers is
fragile. This patch moves the control of those loops to a separate
component that will hopefully be easier to understand and fix when
needed.

In addition this new `work.Loop` component has a `Kick` method that can
be used to wake up the loop when it is sleeping. This is used to wake up
the sync loop automatically before each execution of the watch loop,
which happens when the watch fails, for example. This should make the
loops more reliable, specially in the integration tests.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>