Add hub watcher
This patch changes the hub reconciler so that when a new hub is added it
starts a Kubernetes controller manager with reconcilers to watch the
HyperShift hosted clusters and node pools as well as the Metal3
BareMetalHosts.

The reconciler for the BareMetalHosts will sync the power state of the
host with the fulfillment service.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>