Only update finalizers and status from reconcilers
Currently reconcilers send full updates of objects, which sometimes
results in overwriting changes made by users to the spec of the object.
To reduce the chances of that happening, this patch changes all
reconcilers so that they will use an update mask to ensure that they
only update the `metadata.finalizers` and the `status` of the objects.

This issue was discovered by Elad Tabak, and he is working on a more
general solution in https://github.com/innabox/fulfillment-service/pull/214.
Meanwhile, I am copying part of his work to make the integration tests
less flaky.

Signed-off-by: Elad Tabak <etabak@redhat.com>
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>