Add option to deploy with Kustomize
This patch adds to the integration tests a new `IT_DEPLOY_MODE`
environment variable that is used to define how to deploy the service.
The value can be `helm` or `kustomize`. When the value is `helm` the
service will be deployed using the Helm chart. When the value is
Kustomize the service will be deployed using the Kustomize manifests. In
both cases the tests will be the same, the only difference is how the
service is deployed. The default is `helm`.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>