Support namespace other than _innabox_
Currently the Rego authorization rules assume that the namespace name is
`innabox`. If it is different then the authorization for the `admin`,
`client` and `controller` service accounts will not work because the
`innabox` name is hard-coded. To avoid that this patch changes the
template that generates those rules so that it uses the Helm release
namespace name instead.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>