Better status reporting for controller
This patch changes the controller so that it reports SERVING/NOT_SERVING
depending on the status of the watches: if all the watches are working
correctly then it reports `SERVING`, otherwise it reports `NOT_SERVING`.
In order to do that the patch introduces a new `health.Reporter`
interface and a new `health.Aggreagor` component that aggregates the
satus reported by multiple other components.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>